+++ paks/cli/build.py
@@ -16,7 +16,13 @@ def main(args, parser, extra, subparser):
     # Finally, add the repository and install packages
     if repo:
         cli.add_repository(repo.repo_dir)
-    cache = cli.build(args.packages, args.cache_dir, args.key)
+    cache = cli.build(
+        args.packages,
+        cache_dir=args.cache_dir,
+        key=args.key,
+        registry=args.registry,
+        tag=args.tag,
+    )
 
     # We cannot have both
     if args.push and args.push_trusted:
@@ -24,8 +30,14 @@ def main(args, parser, extra, subparser):
 
     # Do we want to push to a build cache?
     if args.push or args.push_trusted:
-        cache.push(args.push)
+        cache.push(
+            args.push,
+        )
 
-        # By default, we clean up the build cache
-        if not args.no_cleanup:
+        # By default, we clean up the build cache, unless a custom cache is used
+        if not args.no_cleanup and not args.cache_dir:
+            cache.remove()
+
+        # Otherwse, require --no-cleanup and --force
+        elif not args.no_cleanup and args.cache_dir and args.force:
             cache.remove()